requirements.txt
mysql.connector
dnspython
pymongo
PyMySQL
pyrogram
python-dotenv
TgCrypto
motor